requirements.txt
# Produza ERP - Backend Dependencies
# Para deploy no Railway

# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.29
asyncpg==0.29.0
psycopg[binary,pool]==3.3.3
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.6.4
pydantic-core==2.16.3
email-validator==2.1.1

# Environment
python-dotenv==1.0.1

# Utils
greenlet==3.0.3
httpx==0.28.1
